fix compiler warnings

+++ lib/linux_systemutil.hpp
@@ -16,9 +16,9 @@ public:
     static int killProcessById(int pid, std::string procName);
     static int getProcIdByName(std::string procName);
     static bool setAppAsDaemon();
-    static long long userAvailableFreeSpace();
+    static uint64_t userAvailableFreeSpace();
     static long getCPUtemp(void);
-    static long getFreeDiskSpace(std::string absoluteFilePath);
+    static uint64_t getFreeDiskSpace(std::string absoluteFilePath);
     static uint64_t getSysUpTime();
     static uint32_t getNumOfThreadsByThisProcess();
     static uint32_t getNumOfThreadsByPID(int Pid);
